requirement.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.14
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0
django-allauth==0.51.0
django-defender==0.9.5
django-extensions==3.2.1
django-jazzmin==2.5.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
django-user-agents==0.4.0
django-user-visit==0.5.1
djangorestframework==3.14.0
docutils==0.19
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jsonfield==3.1.0
msgpack==1.0.4
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
types-cryptography==3.3.23
typing_extensions==4.4.0
tzdata==2022.2
ua-parser==0.16.1
urllib3==1.26.12
user-agents==2.2.0
wrapt==1.14.1
zope.interface==5.5.0
